% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildGraphc}
\alias{buildGraphc}
\title{Constructs product graph}
\usage{
buildGraphc(X, Y, V, localdegree, thresh, localthresh)
}
\arguments{
\item{X}{matrix Nx3}

\item{Y}{matrix Nx3}

\item{V}{matrix Nvx2}
}
\description{
Constructs product graph
}

