% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{devsub}
\alias{devsub}
\title{Compute deviation between matrices of coordinates X index I and Y index J}
\usage{
devsub(X, I, Y, J)
}
\arguments{
\item{X}{matrix Nx3}

\item{I}{vector M}

\item{Y}{matrix Nx3}

\item{J}{vector M}
}
\description{
Compute deviation between matrices of coordinates X index I and Y index J
}

