% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNeighborsc}
\alias{getNeighborsc}
\title{Get set of neighbors that could be added to the clique C}
\usage{
getNeighborsc(C, V, X, Y, thresh, nbnei)
}
\description{
Get set of neighbors that could be added to the clique C
}

