% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchSearch.R
\name{patchsearch_rmsd}
\alias{patchsearch_rmsd}
\title{Compute rmsd between X and Y}
\usage{
patchsearch_rmsd(X, Y)
}
\arguments{
\item{X}{matrix Nx3}

\item{Y}{matrix Nx3}
}
\description{
Compute rmsd between X and Y
}

