% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchSearch.R
\name{psearch}
\alias{psearch}
\title{search for matchings between query pdb and a surface pdb files}
\usage{
psearch(query, target, mindist = patchsearch_MINDIST,
  maxdist = patchsearch_MAXDIST, minclique = patchsearch_MINCLIQUE,
  mergeclique = patchsearch_MERGECLIQUE, bcmin = patchsearch_BCMIN,
  seqhom = patchsearch_SEQHOM, enrich = TRUE, merge = TRUE,
  ptypes = patchsearch_ATYPES, locdegree = patchsearch_LOCDEGREE,
  locsize = patchsearch_LOCSIZE)
}
\description{
search for matchings between query pdb and a surface pdb files
}

