% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchSearch.R
\name{read_surface}
\alias{read_surface}
\title{read PDB file containing an extracted protein surface or patch}
\usage{
read_surface(pdbfile)
}
\arguments{
\item{file}{pdb file}
}
\description{
read PDB file containing an extracted protein surface or patch
}

