% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmsd}
\alias{rmsd}
\title{Compute RMSD between matrices of coordinates X and Y}
\usage{
rmsd(X, Y)
}
\arguments{
\item{X}{matrix Nx3}

\item{Y}{matrix Nx3}
}
\description{
Compute RMSD between matrices of coordinates X and Y
}

