% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchSearch.R
\name{surf_extract}
\alias{surf_extract}
\title{extract a surface from a pdb file given a list of chains}
\usage{
surf_extract(pdb, chains, minacc = MINACC, add.calpha = TRUE,
  method = "naccess")
}
\description{
extract a surface from a pdb file given a list of chains
}

